Get kardl Package Options

This function retrieves the current settings of the kardl package. Users can specify option names to get their values or call the function without arguments to retrieve all current settings.

Usage

kardl_get(...)

Arguments

...

Option names to retrieve. If no arguments are provided, all options will be returned.

Value

If no arguments are provided, returns all options as a list. If specific option names are provided, returns their values.

See also

Examples

# Get all options
kardl_get()
#> $batch
#> [1] "1/1"
#> 
#> $long_coef
#> [1] "L{lag}.{var_name}"
#> 
#> $maxlag
#> [1] 2
#> 
#> $different_asym_lag
#> [1] TRUE
#> 
#> $asym_suffix
#> [1] "_PP" "_NN"
#> 
#> $asym_prefix
#> [1] "asyP_" "asyN_"
#> 
#> $formula
#> CPI ~ ER + PPI + Asymmetr(ER) + deterministic(covid) + trend
#> <environment: 0x5e0fbd550a98>
#> 
#> $criterion
#> [1] "HQ"
#> 
#> $short_coef
#> [1] "L{lag}.d.{var_name}"
#> 
#> $data
#>                   ER         CPI      PPI covid
#> Jan 1985 -7.69807437 -4.89530369 3.514587     0
#> Feb 1985 -7.63486263 -4.86283642 3.393818     0
#> Mar 1985 -7.61890351 -4.81612238 3.492322     0
#> Apr 1985 -7.57664897 -4.80784226 3.452126     0
#> May 1985 -7.54579136 -4.78411986 3.548855     0
#> Jun 1985 -7.53190825 -4.79175440 3.275258     0
#> Jul 1985 -7.53863213 -4.77760957 3.440338     0
#> Aug 1985 -7.53155365 -4.75206920 3.412280     0
#> Sep 1985 -7.51483486 -4.70567517 3.465705     0
#> Oct 1985 -7.50278713 -4.64493414 3.627260     0
#> Nov 1985 -7.49115160 -4.59957878 3.662690     0
#> Dec 1985 -7.45791962 -4.58356283 3.598995     0
#> Jan 1986 -7.44072427 -4.55206838 3.497935     0
#> Feb 1986 -7.42374421 -4.53483554 3.386017     0
#> Mar 1986 -7.31597167 -4.52187957 3.515428     0
#> Apr 1986 -7.32605490 -4.51810065 3.455887     0
#> May 1986 -7.28633993 -4.49928274 3.497935     0
#> Jun 1986 -7.29032695 -4.47567228 3.474124     0
#> Jul 1986 -7.29359425 -4.45711315 3.543920     0
#> Aug 1986 -7.28855460 -4.44844052 3.521192     0
#> Sep 1986 -7.26604581 -4.42567493 3.593245     0
#> Oct 1986 -7.21219701 -4.35484123 3.704471     0
#> Nov 1986 -7.19935836 -4.33232209 3.670404     0
#> Dec 1986 -7.18510426 -4.31605332 3.660454     0
#> Jan 1987 -7.18808445 -4.28706058 3.582492     0
#> Feb 1987 -7.17405427 -4.26042112 3.538286     0
#> Mar 1987 -7.15886116 -4.22373745 3.587883     0
#> Apr 1987 -7.13448024 -4.20300649 3.571623     0
#> May 1987 -7.10492655 -4.15484354 3.566143     0
#> Jun 1987 -7.06742533 -4.15589312 3.665442     0
#> Jul 1987 -7.03553752 -4.13693065 3.577072     0
#> Aug 1987 -7.01641568 -4.11993146 3.587883     0
#> Sep 1987 -6.97632324 -4.09110710 3.694855     0
#> Oct 1987 -6.96241212 -4.04421289 3.795859     0
#> Nov 1987 -6.93798771 -3.98362347 3.795859     0
#> Dec 1987 -6.88707069 -3.87746751 3.787087     0
#> Jan 1988 -6.79402489 -3.81934572 3.704471     0
#> Feb 1988 -6.74275792 -3.78267038 3.685147     0
#> Mar 1988 -6.70583536 -3.72874899 3.713995     0
#> Apr 1988 -6.66915579 -3.66827876 3.660454     0
#> May 1988 -6.62755110 -3.63427428 3.624827     0
#> Jun 1988 -6.57894086 -3.62535890 3.650404     0
#> Jul 1988 -6.54006127 -3.60950893 3.603884     0
#> Aug 1988 -6.47938778 -3.58591054 3.675343     0
#> Sep 1988 -6.40658007 -3.52843619 3.685147     0
#> Oct 1988 -6.38699913 -3.47408637 3.773783     0
#> Nov 1988 -6.33213117 -3.42205585 3.728113     0
#> Dec 1988 -6.31175797 -3.39772846 3.755764     0
#> Jan 1989 -6.27484654 -3.33383806 3.640251     0
#> Feb 1989 -6.25046328 -3.28719743 3.635136     0
#> Mar 1989 -6.20096657 -3.25727002 3.728113     0
#> Apr 1989 -6.17686925 -3.19688633 3.609161     0
#> May 1989 -6.16739060 -3.16688598 3.640251     0
#> Jun 1989 -6.14684665 -3.14615218 3.723429     0
#> Jul 1989 -6.14096738 -3.11619150 3.645340     0
#> Aug 1989 -6.11259769 -3.08159075 3.728113     0
#> Sep 1989 -6.10183314 -3.02678247 3.795859     0
#> Oct 1989 -6.07541153 -2.96682772 3.813175     0
#> Nov 1989 -6.07150825 -2.92425533 3.817458     0
#> Dec 1989 -6.06891163 -2.90133241 3.791482     0
#> Jan 1990 -6.05348890 -2.86361303 3.742034     0
#> Feb 1990 -6.02788375 -2.82036308 3.718723     0
#> Mar 1990 -5.99687516 -2.76981142 3.855198     0
#> Apr 1990 -5.98272287 -2.70552453 3.690013     0
#> May 1990 -5.95478167 -2.67532448 3.782672     0
#> Jun 1990 -5.93042966 -2.66033444 3.795859     0
#> Jul 1990 -5.92480453 -2.66962021 3.660454     0
#> Aug 1990 -5.91735572 -2.64445584 3.769308     0
#> Sep 1990 -5.90116325 -2.56105672 3.895526     0
#> Oct 1990 -5.88594484 -2.49491146 3.967939     0
#> Nov 1990 -5.87308080 -2.44606718 3.949385     0
#> Dec 1990 -5.83272897 -2.42875034 3.867468     0
#> Jan 1991 -5.79427080 -2.38125528 3.665442     0
#> Feb 1991 -5.70368148 -2.32844576 3.751208     0
#> Mar 1991 -5.59826367 -2.28569711 3.883598     0
#> Apr 1991 -5.53225901 -2.22219560 3.713995     0
#> May 1991 -5.50781939 -2.18942959 3.907314     0
#> Jun 1991 -5.44111210 -2.16029710 3.704471     0
#> Jul 1991 -5.42624165 -2.14725718 3.838600     0
#> Aug 1991 -5.38248189 -2.10821801 3.838600     0
#> Sep 1991 -5.35525392 -2.04912896 3.911213     0
#> Oct 1991 -5.31260495 -1.98527143 3.975265     0
#> Nov 1991 -5.28482283 -1.93414998 3.956848     0
#> Dec 1991 -5.28245977 -1.89145408 3.911213     0
#> Jan 1992 -5.20455202 -1.80186008 3.855198     0
#> Feb 1992 -5.13736927 -1.75273787 3.808874     0
#> Mar 1992 -5.07553388 -1.70522186 3.875565     0
#> Apr 1992 -5.02062503 -1.66835252 3.773783     0
#> May 1992 -4.98684504 -1.65961282 3.891566     0
#> Jun 1992 -4.97984618 -1.65473553 3.787087     0
#> Jul 1992 -4.95247064 -1.64162704 3.883598     0
#> Aug 1992 -4.95360205 -1.60416401 3.842776     0
#> Sep 1992 -4.91650718 -1.53232663 3.971608     0
#> Oct 1992 -4.84773231 -1.45945470 4.018128     0
#> Nov 1992 -4.79274150 -1.41183135 3.975265     0
#> Dec 1992 -4.76013770 -1.38483802 3.949385     0
#> Jan 1993 -4.73037406 -1.33330249 3.887590     0
#> Feb 1993 -4.68911717 -1.29430706 3.851074     0
#> Mar 1993 -4.66062687 -1.24773220 3.842776     0
#> Apr 1993 -4.63983099 -1.20480667 3.915096     0
#> May 1993 -4.59187348 -1.15866439 3.922818     0
#> Jun 1993 -4.52164555 -1.14050163 3.863395     0
#> Jul 1993 -4.46163710 -1.09283679 3.967939     0
#> Aug 1993 -4.45604095 -1.06648175 3.871525     0
#> Sep 1993 -4.41504664 -1.01207468 4.028563     0
#> Oct 1993 -4.34434480 -0.94543224 4.055869     0
#> Nov 1993 -4.28764416 -0.88345055 4.025097     0
#> Dec 1993 -4.23550360 -0.84790038 4.082448     0
#> Jan 1994 -4.06163965 -0.80475512 3.967939     0
#> Feb 1994 -4.00201828 -0.74167124 3.813175     0
#> Mar 1994 -3.80947454 -0.68507903 3.863395     0
#> Apr 1994 -3.37784694 -0.47494947 3.804554     0
#> May 1994 -3.44947806 -0.39567934 3.742034     0
#> Jun 1994 -3.46750451 -0.37630094 3.773783     0
#> Jul 1994 -3.47222346 -0.34546796 3.791482     0
#> Aug 1994 -3.41173299 -0.31746039 3.867468     0
#> Sep 1994 -3.37925293 -0.25739574 3.953124     0
#> Oct 1994 -3.32816070 -0.18726740 3.993348     0
#> Nov 1994 -3.31429976 -0.12565983 3.996926     0
#> Dec 1994 -3.25124407 -0.05803060 3.960559     0
#> Jan 1995 -3.20809663  0.01012958 3.887590     0
#> Feb 1995 -3.18511470  0.05770927 3.791482     0
#> Mar 1995 -3.17411759  0.10187643 3.842776     0
#> Apr 1995 -3.16122031  0.15847827 3.934291     0
#> May 1995 -3.15367346  0.19107702 3.871525     0
#> Jun 1995 -3.12554488  0.21652393 3.989758     0
#> Jul 1995 -3.10124836  0.24677074 3.971608     0
#> Aug 1995 -3.03732540  0.28919494 3.993348     0
#> Sep 1995 -3.01581254  0.36322909 4.011111     0
#> Oct 1995 -2.97371641  0.42447717 4.014626     0
#> Nov 1995 -2.90773971  0.47056436 4.098709     0
#> Dec 1995 -2.81926113  0.50755042 4.105140     0
#> Jan 1996 -2.77003199  0.58733626 4.000491     0
#> Feb 1996 -2.71883566  0.63143605 3.808874     0
#> Mar 1996 -2.64858154  0.68596901 3.964255     0
#> Apr 1996 -2.59801040  0.75085672 3.918965     0
#> May 1996 -2.54816599  0.79502386 4.014626     0
#> Jun 1996 -2.50733632  0.82006771 4.007584     0
#> Jul 1996 -2.48887853  0.84121655 4.021619     0
#> Aug 1996 -2.45320452  0.89164236 4.004044     0
#> Sep 1996 -2.39078809  0.94720457 4.032018     0
#> Oct 1996 -2.34473610  1.00999383 4.158194     0
#> Nov 1996 -2.28695784  1.06046601 4.155150     0
#> Dec 1996 -2.22770956  1.09399139 4.105140     0
#> Jan 1997 -2.15753282  1.15107248 3.996926     0
#> Feb 1997 -2.10549808  1.20610875 3.899471     0
#> Mar 1997 -2.06077682  1.25866719 4.105140     0
#> Apr 1997 -2.00281389  1.32273802 4.000491     0
#> May 1997 -1.96572008  1.36858608 4.136688     0
#> Jun 1997 -1.91023900  1.39693833 4.117879     0
#> Jul 1997 -1.83856810  1.45765369 4.130458     0
#> Aug 1997 -1.79691272  1.51778459 4.127328     0
#> Sep 1997 -1.75082868  1.58833645 4.208575     0
#> Oct 1997 -1.71105544  1.66832470 4.217209     0
#> Nov 1997 -1.64049008  1.73250343 4.170279     0
#> Dec 1997 -1.58179843  1.78260050 4.185183     0
#> Jan 1998 -1.53849029  1.85230056 4.021619     0
#> Feb 1998 -1.47381496  1.89551331 4.052496     0
#> Mar 1998 -1.41792952  1.93763368 4.170279     0
#> Apr 1998 -1.39022206  1.98321874 4.021619     0
#> May 1998 -1.36041221  2.01763844 4.179248     0
#> Jun 1998 -1.32606511  2.04168539 4.139789     0
#> Jul 1998 -1.30966671  2.07470430 4.149034     0
#> Aug 1998 -1.28487560  2.11347666 4.133578     0
#> Sep 1998 -1.28418914  2.17849648 4.208575     0
#> Oct 1998 -1.25256799  2.23725403 4.185183     0
#> Nov 1998 -1.19372549  2.27928606 4.136688     0
#> Dec 1998 -1.15688568  2.31161539 4.079164     0
#> Jan 1999 -1.10442012  2.35852612 3.934291     0
#> Feb 1999 -1.04537773  2.38976457 3.993348     0
#> Mar 1999 -0.99894163  2.42950563 4.055869     0
#> Apr 1999 -0.94337677  2.47703005 4.072564     0
#> May 1999 -0.90234345  2.50601406 4.139789     0
#> Jun 1999 -0.86666758  2.53802449 4.152097     0
#> Jul 1999 -0.85106260  2.57550103 4.133578     0
#> Aug 1999 -0.81188342  2.61670085 4.025097     0
#> Sep 1999 -0.77762287  2.67483768 4.136688     0
#> Oct 1999 -0.73716177  2.73618179 4.111530     0
#> Nov 1999 -0.66376121  2.77734163 4.136688     0
#> Dec 1999 -0.61359690  2.83511308 4.130458     0
#> Jan 2000 -0.58119981  2.88253376 3.964255     0
#> Feb 2000 -0.55610894  2.91889996 4.055869     0
#> Mar 2000 -0.52851788  2.94768130 4.055869     0
#> Apr 2000 -0.49302827  2.97065762 4.117879     0
#> May 2000 -0.48414475  2.99259620 4.185183     0
#> Jun 2000 -0.47870054  2.99935435 4.194019     0
#> Jul 2000 -0.45091341  3.02155790 4.176267     0
#> Aug 2000 -0.42308340  3.04304767 4.199867     0
#> Sep 2000 -0.40422937  3.07333410 4.214339     0
#> Oct 2000 -0.37931210  3.10385117 4.231434     0
#> Nov 2000 -0.38016878  3.14029768 4.237067     0
#> Dec 2000 -0.39543805  3.16465874 4.079164     0
#> Jan 2001 -0.38851655  3.18941215 4.049112     0
#> Feb 2001 -0.08023548  3.20724658 4.014626     0
#> Mar 2001  0.02276001  3.26619106 3.986154     0
#> Apr 2001  0.13141152  3.36453863 4.000491     0
#> May 2001  0.18983989  3.41384022 4.072564     0
#> Jun 2001  0.22776804  3.44469203 4.072564     0
#> Jul 2001  0.28239236  3.46835159 4.065919     0
#> Aug 2001  0.30859291  3.49731747 4.092236     0
#> Sep 2001  0.43975638  3.55453431 4.121038     0
#> Oct 2001  0.46450856  3.61348748 4.088984     0
#> Nov 2001  0.39036771  3.65487826 4.082448     0
#> Dec 2001  0.37165114  3.68661245 4.000491     0
#> Jan 2002  0.26887227  3.73844059 4.021619     0
#> Feb 2002  0.33532444  3.75582290 3.960559     0
#> Mar 2002  0.29343052  3.76765152 4.155150     0
#> Apr 2002  0.28900320  3.78799709 4.145962     0
#> May 2002  0.36566898  3.79377285 4.185183     0
#> Jun 2002  0.45293821  3.79959297 4.149034     0
#> Jul 2002  0.52231319  3.81390501 4.182220     0
#> Aug 2002  0.48566596  3.83551272 4.158194     0
#> Sep 2002  0.50346017  3.86969518 4.222923     0
#> Oct 2002  0.51073822  3.90203370 4.214339     0
#> Nov 2002  0.43116014  3.93073584 4.176267     0
#> Dec 2002  0.49694919  3.94704240 4.139789     0
#> Jan 2003  0.49437546  3.97259487 4.176267     0
#> Feb 2003  0.46524859  3.98788313 4.000491     0
#> Mar 2003  0.53307995  4.00733319 4.222923     0
#> Apr 2003  0.45174928  4.01717049 4.167272     0
#> May 2003  0.36333879  4.02671207 4.214339     0
#> Jun 2003  0.34432844  4.02751143 4.234254     0
#> Jul 2003  0.34728606  4.02561191 4.278339     0
#> Aug 2003  0.33371773  4.02721175 4.248241     0
#> Sep 2003  0.32765959  4.04060946 4.328283     0
#> Oct 2003  0.39371467  4.04983335 4.315380     0
#> Nov 2003  0.37761041  4.06245120 4.194019     0
#> Dec 2003  0.33406792  4.06668605 4.323142     0
#> Jan 2004  0.29283788  4.07329115 4.228605     0
#> Feb 2004  0.28102974  4.07843011 4.136688     0
#> Mar 2004  0.27260282  4.08797153 4.341022     0
#> Apr 2004  0.35116167  4.09294223 4.312779     0
#> May 2004  0.40253749  4.09723649 4.370950     0
#> Jun 2004  0.39843714  4.09593149 4.383160     0
#> Jul 2004  0.38266106  4.10067723 4.392822     0
#> Aug 2004  0.40928713  4.10826073 4.328283     0
#> Sep 2004  0.40610490  4.11770558 4.380730     0
#> Oct 2004  0.38777489  4.13990866 4.323142     0
#> Nov 2004  0.35729115  4.15294480 4.278339     0
#> Dec 2004  0.29229641  4.15611159 4.363551     0
#> Jan 2005  0.28720696  4.16162945 4.275640     0
#> Feb 2005  0.24810930  4.16180412 4.302308     0
#> Mar 2005  0.29969710  4.16442055 4.437483     0
#> Apr 2005  0.32768373  4.17153740 4.402391     0
#> May 2005  0.30619711  4.18066281 4.444356     0
#> Jun 2005  0.29035350  4.18169065 4.475819     0
#> Jul 2005  0.28095953  4.17593833 4.453446     0
#> Aug 2005  0.30051192  4.18442639 4.451181     0
#> Sep 2005  0.29672853  4.19461930 4.540081     0
#> Oct 2005  0.29635684  4.21237669 4.562646     0
#> Nov 2005  0.30250911  4.22630989 4.473604     0
#> Dec 2005  0.29643119  4.23047674 4.576745     0
#> Jan 2006  0.27997747  4.23794976 4.286392     0
#> Feb 2006  0.26937807  4.24013238 4.388003     0
#> Mar 2006  0.29635684  4.24287409 4.554499     0
#> Apr 2006  0.27660849  4.25615322 4.499864     0
#> May 2006  0.44708678  4.27473351 4.562646     0
#> Jun 2006  0.45330244  4.27808203 4.578743     0
#> Jul 2006  0.39517905  4.28652025 4.542153     0
#> Aug 2006  0.37245971  4.28211649 4.514869     0
#> Sep 2006  0.40593167  4.29488785 4.584713     0
#> Oct 2006  0.37672264  4.30749817 4.521232     0
#> Nov 2006  0.37108068  4.32032415 4.608243     0
#> Dec 2006  0.34288023  4.32263181 4.600461     0
#> Jan 2007  0.34847139  4.33261969 4.462456     0
#> Feb 2007  0.34811844  4.33688033 4.473604     0
#> Mar 2007  0.32458038  4.34600148 4.614040     0
#> Apr 2007  0.31042156  4.35798719 4.558581     0
#> May 2007  0.27748021  4.36299865 4.635014     0
#> Jun 2007  0.26830810  4.36056765 4.623628     0
#> Jul 2007  0.24502245  4.35323898 4.602412     0
#> Aug 2007  0.25812452  4.35345529 4.592618     0
#> Sep 2007  0.18871772  4.36371253 4.627438     0
#> Oct 2007  0.16079995  4.38167624 4.619804     0
#> Nov 2007  0.16071480  4.40097435 4.686496     0
#> Dec 2007  0.15768728  4.40317200 4.584713     0
#> Jan 2008  0.15768728  4.41116631 4.568713     0
#> Feb 2008  0.17684827  4.42401387 4.560615     0
#> Mar 2008  0.26968357  4.43357564 4.648135     0
#> Apr 2008  0.25309063  4.45020640 4.631233     0
#> May 2008  0.19028973  4.46501859 4.664756     0
#> Jun 2008  0.20012054  4.46140079 4.648135     0
#> Jul 2008  0.14773011  4.46714440 4.644403     0
#> Aug 2008  0.16335108  4.46469611 4.556542     0
#> Sep 2008  0.21070653  4.46920149 4.586695     0
#> Oct 2008  0.44727861  4.49487725 4.554499     0
#> Nov 2008  0.44862038  4.50316726 4.550401     0
#> Dec 2008  0.42228945  4.49906201 4.392822     0
#> Jan 2009  0.49375067  4.50192502 4.325716     0
#> Feb 2009  0.52558615  4.49850089 4.286392     0
#> Mar 2009  0.51414012  4.50947877 4.407142     0
#> Apr 2009  0.45894268  4.50966380 4.423594     0
#> May 2009  0.43061290  4.51605713 4.475819     0
#> Jun 2009  0.42422144  4.51715953 4.542153     0
#> Jul 2009  0.38706350  4.51966603 4.552452     0
#> Aug 2009  0.40479822  4.51666973 4.491187     0
#> Sep 2009  0.39241364  4.52058147 4.484629     0
#> Oct 2009  0.39622253  4.54439140 4.614040     0
#> Nov 2009  0.39615524  4.55699922 4.519116     0
#> Dec 2009  0.39937996  4.56227907 4.608243     0
#> Jan 2010  0.39655890  4.58059948 4.407142     0
#> Feb 2010  0.43165254  4.59497263 4.416576     0
#> Mar 2010  0.41828261  4.60078840 4.586695     0
#> Apr 2010  0.39031763  4.60673846 4.556542     0
#> May 2010  0.45126668  4.60315021 4.600461     0
#> Jun 2010  0.45584132  4.59751766 4.644403     0
#> Jul 2010  0.41088373  4.59270492 4.651852     0
#> Aug 2010  0.42169928  4.59672657 4.614040     0
#> Sep 2010  0.36942334  4.60891876 4.580737     0
#> Oct 2010  0.36015389  4.62706988 4.713022     0
#> Nov 2010  0.40726349  4.62734404 4.614040     0
#> Dec 2010  0.43262622  4.62432414 4.809442     0
#> Jan 2011  0.47187687  4.62843993 4.623628     0
#> Feb 2011  0.46793900  4.63569722 4.592618     0
#> Mar 2011  0.43178242  4.63987485 4.735456     0
#> Apr 2011  0.41726192  4.64850024 4.697191     0
#> May 2011  0.46354532  4.67242013 4.742259     0
#> Jun 2011  0.48217921  4.65800937 4.783759     0
#> Jul 2011  0.51724498  4.65390696 4.770665     0
#> Aug 2011  0.55938719  4.66114142 4.733748     0
#> Sep 2011  0.61826198  4.66863961 4.759066     0
#> Oct 2011  0.55961579  4.70083701 4.863426     0
#> Nov 2011  0.61266896  4.71795821 4.749016     0
#> Dec 2011  0.63840056  4.72370033 4.888631     0
#> Jan 2012  0.56999036  4.72931065 4.682905     0
#> Feb 2012  0.55486165  4.73493890 4.661086     0
#> Mar 2012  0.57503909  4.73906586 4.793468     0
#> Apr 2012  0.56138565  4.75410228 4.750698     0
#> May 2012  0.61239797  4.75197492 4.829835     0
#> Jun 2012  0.59379635  4.74298053 4.822041     0
#> Jul 2012  0.58260660  4.74068252 4.812607     0
#> Aug 2012  0.59633337  4.74624731 4.702496     0
#> Sep 2012  0.58165680  4.75651424 4.832936     0
#> Oct 2012  0.58199213  4.77593418 4.806268     0
#> Nov 2012  0.57756796  4.77970741 4.881284     0
#> Dec 2012  0.57768021  4.78351336 4.884229     0
#> Jan 2013  0.56477389  4.79984044 4.725165     0
#> Feb 2013  0.59144663  4.80283494 4.709526     0
#> Mar 2013  0.59627829  4.80943714 4.818906     0
#> Apr 2013  0.58606296  4.81363250 4.832936     0
#> May 2013  0.62476127  4.81508764 4.885699     0
#> Jun 2013  0.65694989  4.82269259 4.878330     0
#> Jul 2013  0.65536743  4.82579937 4.906048     0
#> Aug 2013  0.72151595  4.82476485 4.716507     0
#> Sep 2013  0.71098710  4.83238615 4.933020     0
#> Oct 2013  0.68841098  4.85022440 4.840646     0
#> Nov 2013  0.70272616  4.85031214 4.974178     0
#> Dec 2013  0.75902855  4.85490764 4.979540     0
#> Jan 2014  0.82487695  4.87449533 4.837569     0
#> Feb 2014  0.80477800  4.87876811 4.788625     0
#> Mar 2014  0.78472312  4.88996131 4.907486     0
#> Apr 2014  0.75136892  4.90332137 4.897378     0
#> May 2014  0.73883731  4.90726582 4.911787     0
#> Jun 2014  0.75391297  4.91036891 4.927402     0
#> Jul 2014  0.73895672  4.91490275 4.872395     0
#> Aug 2014  0.77027025  4.91584815 4.858912     0
#> Sep 2014  0.82459202  4.91724405 4.994138     0
#> Oct 2014  0.79710171  4.93606154 4.891555     0
#> Nov 2014  0.79396489  4.93787171 4.971486     0
#> Dec 2014  0.84197658  4.93344104 5.052979     0
#> Jan 2015  0.88366423  4.94440127 4.828281     0
#> Feb 2015  0.92028275  4.95152298 4.798287     0
#> Mar 2015  0.96032676  4.96330707 4.967435     0
#> Apr 2015  0.98087925  4.97947469 4.963367     0
#> May 2015  0.97873957  4.98508750 4.959282     0
#> Jun 2015  0.98905781  4.97993720 5.015000     0
#> Jul 2015  1.02137121  4.98082310 4.920334     0
#> Aug 2015  1.07064139  4.98481911 4.986202     0
#> Sep 2015  1.11384567  4.99371416 4.941388     0
#> Oct 2015  1.07074422  5.00913941 5.054220     0
#> Nov 2015  1.07231965  5.01577899 5.042991     0
#> Dec 2015  1.06822181  5.01785877 5.127176     0
#> Jan 2016  1.08638787  5.03587462 4.879808     0
#> Feb 2016  1.07565068  5.03565597 4.927402     0
#> Mar 2016  1.04237698  5.03525499 5.032903     0
#> Apr 2016  1.03101131  5.04306341 4.998083     0
#> May 2016  1.08473308  5.04883502 5.039220     0
#> Jun 2016  1.06339953  5.05353584 5.041736     0
#> Jul 2016  1.10368275  5.06510366 4.829835     0
#> Aug 2016  1.08419215  5.06219761 5.023991     0
#> Sep 2016  1.09814551  5.06400603 4.903167     0
#> Oct 2016  1.13224047  5.07828696 5.093155     0
#> Nov 2016  1.23051736  5.08344251 5.099104     0
#> Dec 2016  1.25912838  5.09974338 5.132926     0
#> Jan 2017  1.34439105  5.12405740 4.941388     0
#> Feb 2017  1.27854210  5.13213174 4.933020     0
#> Mar 2017  1.29250552  5.14224032 5.095539     0
#> Apr 2017  1.26796117  5.15522730 5.067776     0
#> May 2017  1.27185105  5.15968007 5.099104     0
#> Jun 2017  1.25570150  5.15700433 5.020148     0
#> Jul 2017  1.26197753  5.15848816 5.088370     0
#> Aug 2017  1.23666262  5.16366427 5.099104     0
#> Sep 2017  1.26843945  5.17015271 5.071441     0
#> Oct 2017  1.32797645  5.19071218 5.200542     0
#> Nov 2017  1.37475302  5.20552084 5.194090     0
#> Dec 2017  1.32847985  5.21235517 5.234265     0
#> Jan 2018  1.33049091  5.22250477 5.059171     0
#> Feb 2018  1.33149493  5.22979484 5.022712     0
#> Mar 2018  1.37433564  5.23968068 5.158960     0
#> Apr 2018  1.40048076  5.25823081 5.109723     0
#> May 2018  1.50127347  5.27432098 5.162307     0
#> Jun 2018  1.52877001  5.30010948 5.040479     0
#> Jul 2018  1.58706960  5.30560576 5.160077     0
#> Aug 2018  1.85817903  5.32830056 4.979540     0
#> Sep 2018  1.79102587  5.38944023 5.108549     0
#> Oct 2018  1.70933754  5.41577648 5.152234     0
#> Nov 2018  1.64278564  5.40124218 5.121392     0
#> Dec 2018  1.66120459  5.39718826 5.129480     0
#> Jan 2019  1.66446572  5.40776984 4.980876     0
#> Feb 2019  1.66681019  5.40937630 4.971486     0
#> Mar 2019  1.72873092  5.41960695 5.112068     0
#> Apr 2019  1.78062268  5.43639628 5.097917     0
#> May 2019  1.77983027  5.44584785 5.161192     0
#> Jun 2019  1.75098088  5.44611382 4.939998     0
#> Jul 2019  1.71846271  5.45958551 5.156723     0
#> Aug 2019  1.76117139  5.46813582 4.962007     0
#> Sep 2019  1.73416567  5.47801965 5.149982     0
#> Oct 2019  1.74771179  5.49784340 5.178872     0
#> Nov 2019  1.74841694  5.50164708 5.161192     0
#> Dec 2019  1.78264304  5.50905241 5.224895     1
#> Jan 2020  1.78791877  5.52246938 5.054220     1
#> Feb 2020  1.82961707  5.52597984 5.078731     1
#> Mar 2020  1.87515808  5.53167761 5.102656     1
#> Apr 2020  1.94376499  5.54018585 4.725165     1
#> May 2020  1.91984481  5.55371537 4.798287     1
#> Jun 2020  1.92400775  5.56498419 5.102656     1
#> Jul 2020  1.93779877  5.57080613 5.149982     1
#> Aug 2020  1.98946213  5.57941249 5.112068     1
#> Sep 2020  2.05605136  5.58909861 5.256807     1
#> Oct 2020  2.11814529  5.61018607 5.270891     1
#> Nov 2020  2.05470690  5.63290602 5.244574     1
#> Dec 2020  1.99430568  5.64532409 5.340350     1
#> Jan 2021  1.99172992  5.66200244 5.134072     1
#> Feb 2021  1.99842805  5.67103999 5.136360     1
#> Mar 2021  2.12028354  5.68173630 5.288714     1
#> Apr 2021  2.10201606  5.69838678 5.231151     1
#> May 2021  2.14679056  5.70723317 5.144329     1
#> Jun 2021  2.16481616  5.72646790 5.315821     1
#> Jul 2021  2.13078202  5.74435332 5.134072     1
#> Aug 2021  2.12618199  5.75548675 5.286749     1
#> Sep 2021  2.18055869  5.76793555 5.339418     1
#> Oct 2021  2.25022281  5.79159074 5.323433     1
#> Nov 2021  2.53930803  5.82610593 5.377849     1
#> Dec 2021  2.56411824  5.95340348 5.462164     1
#> Jan 2022  2.61145486  6.05870139 5.197321     1
#> Feb 2022  2.64277541  6.10566619 5.251728     1
#> Mar 2022  2.68505448  6.15886783 5.365202     1
#> Apr 2022  2.69452582  6.22884484 5.319634     1
#> May 2022  2.79550842  6.25821724 5.249689     1
#> Jun 2022  2.81445317  6.30654938 5.392994     1
#> Jul 2022  2.88535922  6.32992671 5.158960     1
#> Aug 2022  2.90038534  6.34442605 5.325327     1
#> Sep 2022  2.91887553  6.37472111 5.340350     1
#> Oct 2022  2.92341692  6.40955515 5.351472     1
#> Nov 2022  2.92481344  6.43798481 5.360646     1
#> Dec 2022  2.92933336  6.44974225 5.470381     1
#> Jan 2023  2.93409886  6.51411460 5.262867     1
#> Feb 2023  2.93811375  6.54508063 5.160077     1
#> Mar 2023  2.95299177  6.56771728 5.368832     1
#> Apr 2023  2.96755397  6.59129228 5.215436     1
#> May 2023  3.01203368  6.59172294 5.358818     1
#> Jun 2023  3.25216939  6.63017892 5.283795     1
#> Jul 2023  3.29448966  6.72083122 5.337551     1
#> Aug 2023  3.28092061  6.80780487 5.358818     1
#> Sep 2023  3.31059228  6.85424097 5.381433     1
#> Oct 2023  3.34037639  6.88800433 5.406170     1
#> Nov 2023  3.36334320  6.92028852 5.378746     1
#> Dec 2023  3.38319463  6.94914034 5.464636     1
#> Jan 2024  3.41222304  7.01402233 5.277860     1
#> Feb 2024  3.43875325  7.05831849 5.300421     1
#> 
#> $mode
#> [1] "quick"
#> 
# Get specific options
kardl_get("maxlag", "mode")
#> $maxlag
#> [1] 2
#> 
#> $mode
#> [1] "quick"
#> 

# Note: In interactive use, avoid calling kardl_get() directly to prevent
# cluttering the console.

a <- kardl_get()
a$asym_suffix
#> [1] "_PP" "_NN"

# To reset options and then get a default option:
mydefaults <- kardl_reset()
mydefaults$maxlag
#> [1] 4