Get kardl Package Options

This function retrieves the current settings of the kardl package. Users can specify option names to get their values or call the function without arguments to retrieve all current settings.

Usage

kardl_get(...)

Arguments

...

Option names to retrieve. If no arguments are provided, all options will be returned.

Value

If no arguments are provided, returns all options as a list. If specific option names are provided, returns their values.

See also

Examples

# Get all options
kardl_get()
#> $batch
#> [1] "1/1"
#> 
#> $AsymSuffix
#> [1] "_PP" "_NN"
#> 
#> $maxlag
#> [1] 2
#> 
#> $formula
#> CPI ~ ER + PPI + Asymmetr(ER) + deterministic(covid) + trend
#> <environment: 0x5a41105efc48>
#> 
#> $ShortCoef
#> [1] "L{lag}.d.{varName}"
#> 
#> $differentAsymLag
#> [1] TRUE
#> 
#> $criterion
#> [1] "HQ"
#> 
#> $data
#>                  ER         CPI      PPI covid
#> 1985-01 -7.69807437 -4.89530369 3.514587     0
#> 1985-02 -7.63486263 -4.86283642 3.393818     0
#> 1985-03 -7.61890351 -4.81612238 3.492322     0
#> 1985-04 -7.57664897 -4.80784226 3.452126     0
#> 1985-05 -7.54579136 -4.78411986 3.548855     0
#> 1985-06 -7.53190825 -4.79175440 3.275258     0
#> 1985-07 -7.53863213 -4.77760957 3.440338     0
#> 1985-08 -7.53155365 -4.75206920 3.412280     0
#> 1985-09 -7.51483486 -4.70567517 3.465705     0
#> 1985-10 -7.50278713 -4.64493414 3.627260     0
#> 1985-11 -7.49115160 -4.59957878 3.662690     0
#> 1985-12 -7.45791962 -4.58356283 3.598995     0
#> 1986-01 -7.44072427 -4.55206838 3.497935     0
#> 1986-02 -7.42374421 -4.53483554 3.386017     0
#> 1986-03 -7.31597167 -4.52187957 3.515428     0
#> 1986-04 -7.32605490 -4.51810065 3.455887     0
#> 1986-05 -7.28633993 -4.49928274 3.497935     0
#> 1986-06 -7.29032695 -4.47567228 3.474124     0
#> 1986-07 -7.29359425 -4.45711315 3.543920     0
#> 1986-08 -7.28855460 -4.44844052 3.521192     0
#> 1986-09 -7.26604581 -4.42567493 3.593245     0
#> 1986-10 -7.21219701 -4.35484123 3.704471     0
#> 1986-11 -7.19935836 -4.33232209 3.670404     0
#> 1986-12 -7.18510426 -4.31605332 3.660454     0
#> 1987-01 -7.18808445 -4.28706058 3.582492     0
#> 1987-02 -7.17405427 -4.26042112 3.538286     0
#> 1987-03 -7.15886116 -4.22373745 3.587883     0
#> 1987-04 -7.13448024 -4.20300649 3.571623     0
#> 1987-05 -7.10492655 -4.15484354 3.566143     0
#> 1987-06 -7.06742533 -4.15589312 3.665442     0
#> 1987-07 -7.03553752 -4.13693065 3.577072     0
#> 1987-08 -7.01641568 -4.11993146 3.587883     0
#> 1987-09 -6.97632324 -4.09110710 3.694855     0
#> 1987-10 -6.96241212 -4.04421289 3.795859     0
#> 1987-11 -6.93798771 -3.98362347 3.795859     0
#> 1987-12 -6.88707069 -3.87746751 3.787087     0
#> 1988-01 -6.79402489 -3.81934572 3.704471     0
#> 1988-02 -6.74275792 -3.78267038 3.685147     0
#> 1988-03 -6.70583536 -3.72874899 3.713995     0
#> 1988-04 -6.66915579 -3.66827876 3.660454     0
#> 1988-05 -6.62755110 -3.63427428 3.624827     0
#> 1988-06 -6.57894086 -3.62535890 3.650404     0
#> 1988-07 -6.54006127 -3.60950893 3.603884     0
#> 1988-08 -6.47938778 -3.58591054 3.675343     0
#> 1988-09 -6.40658007 -3.52843619 3.685147     0
#> 1988-10 -6.38699913 -3.47408637 3.773783     0
#> 1988-11 -6.33213117 -3.42205585 3.728113     0
#> 1988-12 -6.31175797 -3.39772846 3.755764     0
#> 1989-01 -6.27484654 -3.33383806 3.640251     0
#> 1989-02 -6.25046328 -3.28719743 3.635136     0
#> 1989-03 -6.20096657 -3.25727002 3.728113     0
#> 1989-04 -6.17686925 -3.19688633 3.609161     0
#> 1989-05 -6.16739060 -3.16688598 3.640251     0
#> 1989-06 -6.14684665 -3.14615218 3.723429     0
#> 1989-07 -6.14096738 -3.11619150 3.645340     0
#> 1989-08 -6.11259769 -3.08159075 3.728113     0
#> 1989-09 -6.10183314 -3.02678247 3.795859     0
#> 1989-10 -6.07541153 -2.96682772 3.813175     0
#> 1989-11 -6.07150825 -2.92425533 3.817458     0
#> 1989-12 -6.06891163 -2.90133241 3.791482     0
#> 1990-01 -6.05348890 -2.86361303 3.742034     0
#> 1990-02 -6.02788375 -2.82036308 3.718723     0
#> 1990-03 -5.99687516 -2.76981142 3.855198     0
#> 1990-04 -5.98272287 -2.70552453 3.690013     0
#> 1990-05 -5.95478167 -2.67532448 3.782672     0
#> 1990-06 -5.93042966 -2.66033444 3.795859     0
#> 1990-07 -5.92480453 -2.66962021 3.660454     0
#> 1990-08 -5.91735572 -2.64445584 3.769308     0
#> 1990-09 -5.90116325 -2.56105672 3.895526     0
#> 1990-10 -5.88594484 -2.49491146 3.967939     0
#> 1990-11 -5.87308080 -2.44606718 3.949385     0
#> 1990-12 -5.83272897 -2.42875034 3.867468     0
#> 1991-01 -5.79427080 -2.38125528 3.665442     0
#> 1991-02 -5.70368148 -2.32844576 3.751208     0
#> 1991-03 -5.59826367 -2.28569711 3.883598     0
#> 1991-04 -5.53225901 -2.22219560 3.713995     0
#> 1991-05 -5.50781939 -2.18942959 3.907314     0
#> 1991-06 -5.44111210 -2.16029710 3.704471     0
#> 1991-07 -5.42624165 -2.14725718 3.838600     0
#> 1991-08 -5.38248189 -2.10821801 3.838600     0
#> 1991-09 -5.35525392 -2.04912896 3.911213     0
#> 1991-10 -5.31260495 -1.98527143 3.975265     0
#> 1991-11 -5.28482283 -1.93414998 3.956848     0
#> 1991-12 -5.28245977 -1.89145408 3.911213     0
#> 1992-01 -5.20455202 -1.80186008 3.855198     0
#> 1992-02 -5.13736927 -1.75273787 3.808874     0
#> 1992-03 -5.07553388 -1.70522186 3.875565     0
#> 1992-04 -5.02062503 -1.66835252 3.773783     0
#> 1992-05 -4.98684504 -1.65961282 3.891566     0
#> 1992-06 -4.97984618 -1.65473553 3.787087     0
#> 1992-07 -4.95247064 -1.64162704 3.883598     0
#> 1992-08 -4.95360205 -1.60416401 3.842776     0
#> 1992-09 -4.91650718 -1.53232663 3.971608     0
#> 1992-10 -4.84773231 -1.45945470 4.018128     0
#> 1992-11 -4.79274150 -1.41183135 3.975265     0
#> 1992-12 -4.76013770 -1.38483802 3.949385     0
#> 1993-01 -4.73037406 -1.33330249 3.887590     0
#> 1993-02 -4.68911717 -1.29430706 3.851074     0
#> 1993-03 -4.66062687 -1.24773220 3.842776     0
#> 1993-04 -4.63983099 -1.20480667 3.915096     0
#> 1993-05 -4.59187348 -1.15866439 3.922818     0
#> 1993-06 -4.52164555 -1.14050163 3.863395     0
#> 1993-07 -4.46163710 -1.09283679 3.967939     0
#> 1993-08 -4.45604095 -1.06648175 3.871525     0
#> 1993-09 -4.41504664 -1.01207468 4.028563     0
#> 1993-10 -4.34434480 -0.94543224 4.055869     0
#> 1993-11 -4.28764416 -0.88345055 4.025097     0
#> 1993-12 -4.23550360 -0.84790038 4.082448     0
#> 1994-01 -4.06163965 -0.80475512 3.967939     0
#> 1994-02 -4.00201828 -0.74167124 3.813175     0
#> 1994-03 -3.80947454 -0.68507903 3.863395     0
#> 1994-04 -3.37784694 -0.47494947 3.804554     0
#> 1994-05 -3.44947806 -0.39567934 3.742034     0
#> 1994-06 -3.46750451 -0.37630094 3.773783     0
#> 1994-07 -3.47222346 -0.34546796 3.791482     0
#> 1994-08 -3.41173299 -0.31746039 3.867468     0
#> 1994-09 -3.37925293 -0.25739574 3.953124     0
#> 1994-10 -3.32816070 -0.18726740 3.993348     0
#> 1994-11 -3.31429976 -0.12565983 3.996926     0
#> 1994-12 -3.25124407 -0.05803060 3.960559     0
#> 1995-01 -3.20809663  0.01012958 3.887590     0
#> 1995-02 -3.18511470  0.05770927 3.791482     0
#> 1995-03 -3.17411759  0.10187643 3.842776     0
#> 1995-04 -3.16122031  0.15847827 3.934291     0
#> 1995-05 -3.15367346  0.19107702 3.871525     0
#> 1995-06 -3.12554488  0.21652393 3.989758     0
#> 1995-07 -3.10124836  0.24677074 3.971608     0
#> 1995-08 -3.03732540  0.28919494 3.993348     0
#> 1995-09 -3.01581254  0.36322909 4.011111     0
#> 1995-10 -2.97371641  0.42447717 4.014626     0
#> 1995-11 -2.90773971  0.47056436 4.098709     0
#> 1995-12 -2.81926113  0.50755042 4.105140     0
#> 1996-01 -2.77003199  0.58733626 4.000491     0
#> 1996-02 -2.71883566  0.63143605 3.808874     0
#> 1996-03 -2.64858154  0.68596901 3.964255     0
#> 1996-04 -2.59801040  0.75085672 3.918965     0
#> 1996-05 -2.54816599  0.79502386 4.014626     0
#> 1996-06 -2.50733632  0.82006771 4.007584     0
#> 1996-07 -2.48887853  0.84121655 4.021619     0
#> 1996-08 -2.45320452  0.89164236 4.004044     0
#> 1996-09 -2.39078809  0.94720457 4.032018     0
#> 1996-10 -2.34473610  1.00999383 4.158194     0
#> 1996-11 -2.28695784  1.06046601 4.155150     0
#> 1996-12 -2.22770956  1.09399139 4.105140     0
#> 1997-01 -2.15753282  1.15107248 3.996926     0
#> 1997-02 -2.10549808  1.20610875 3.899471     0
#> 1997-03 -2.06077682  1.25866719 4.105140     0
#> 1997-04 -2.00281389  1.32273802 4.000491     0
#> 1997-05 -1.96572008  1.36858608 4.136688     0
#> 1997-06 -1.91023900  1.39693833 4.117879     0
#> 1997-07 -1.83856810  1.45765369 4.130458     0
#> 1997-08 -1.79691272  1.51778459 4.127328     0
#> 1997-09 -1.75082868  1.58833645 4.208575     0
#> 1997-10 -1.71105544  1.66832470 4.217209     0
#> 1997-11 -1.64049008  1.73250343 4.170279     0
#> 1997-12 -1.58179843  1.78260050 4.185183     0
#> 1998-01 -1.53849029  1.85230056 4.021619     0
#> 1998-02 -1.47381496  1.89551331 4.052496     0
#> 1998-03 -1.41792952  1.93763368 4.170279     0
#> 1998-04 -1.39022206  1.98321874 4.021619     0
#> 1998-05 -1.36041221  2.01763844 4.179248     0
#> 1998-06 -1.32606511  2.04168539 4.139789     0
#> 1998-07 -1.30966671  2.07470430 4.149034     0
#> 1998-08 -1.28487560  2.11347666 4.133578     0
#> 1998-09 -1.28418914  2.17849648 4.208575     0
#> 1998-10 -1.25256799  2.23725403 4.185183     0
#> 1998-11 -1.19372549  2.27928606 4.136688     0
#> 1998-12 -1.15688568  2.31161539 4.079164     0
#> 1999-01 -1.10442012  2.35852612 3.934291     0
#> 1999-02 -1.04537773  2.38976457 3.993348     0
#> 1999-03 -0.99894163  2.42950563 4.055869     0
#> 1999-04 -0.94337677  2.47703005 4.072564     0
#> 1999-05 -0.90234345  2.50601406 4.139789     0
#> 1999-06 -0.86666758  2.53802449 4.152097     0
#> 1999-07 -0.85106260  2.57550103 4.133578     0
#> 1999-08 -0.81188342  2.61670085 4.025097     0
#> 1999-09 -0.77762287  2.67483768 4.136688     0
#> 1999-10 -0.73716177  2.73618179 4.111530     0
#> 1999-11 -0.66376121  2.77734163 4.136688     0
#> 1999-12 -0.61359690  2.83511308 4.130458     0
#> 2000-01 -0.58119981  2.88253376 3.964255     0
#> 2000-02 -0.55610894  2.91889996 4.055869     0
#> 2000-03 -0.52851788  2.94768130 4.055869     0
#> 2000-04 -0.49302827  2.97065762 4.117879     0
#> 2000-05 -0.48414475  2.99259620 4.185183     0
#> 2000-06 -0.47870054  2.99935435 4.194019     0
#> 2000-07 -0.45091341  3.02155790 4.176267     0
#> 2000-08 -0.42308340  3.04304767 4.199867     0
#> 2000-09 -0.40422937  3.07333410 4.214339     0
#> 2000-10 -0.37931210  3.10385117 4.231434     0
#> 2000-11 -0.38016878  3.14029768 4.237067     0
#> 2000-12 -0.39543805  3.16465874 4.079164     0
#> 2001-01 -0.38851655  3.18941215 4.049112     0
#> 2001-02 -0.08023548  3.20724658 4.014626     0
#> 2001-03  0.02276001  3.26619106 3.986154     0
#> 2001-04  0.13141152  3.36453863 4.000491     0
#> 2001-05  0.18983989  3.41384022 4.072564     0
#> 2001-06  0.22776804  3.44469203 4.072564     0
#> 2001-07  0.28239236  3.46835159 4.065919     0
#> 2001-08  0.30859291  3.49731747 4.092236     0
#> 2001-09  0.43975638  3.55453431 4.121038     0
#> 2001-10  0.46450856  3.61348748 4.088984     0
#> 2001-11  0.39036771  3.65487826 4.082448     0
#> 2001-12  0.37165114  3.68661245 4.000491     0
#> 2002-01  0.26887227  3.73844059 4.021619     0
#> 2002-02  0.33532444  3.75582290 3.960559     0
#> 2002-03  0.29343052  3.76765152 4.155150     0
#> 2002-04  0.28900320  3.78799709 4.145962     0
#> 2002-05  0.36566898  3.79377285 4.185183     0
#> 2002-06  0.45293821  3.79959297 4.149034     0
#> 2002-07  0.52231319  3.81390501 4.182220     0
#> 2002-08  0.48566596  3.83551272 4.158194     0
#> 2002-09  0.50346017  3.86969518 4.222923     0
#> 2002-10  0.51073822  3.90203370 4.214339     0
#> 2002-11  0.43116014  3.93073584 4.176267     0
#> 2002-12  0.49694919  3.94704240 4.139789     0
#> 2003-01  0.49437546  3.97259487 4.176267     0
#> 2003-02  0.46524859  3.98788313 4.000491     0
#> 2003-03  0.53307995  4.00733319 4.222923     0
#> 2003-04  0.45174928  4.01717049 4.167272     0
#> 2003-05  0.36333879  4.02671207 4.214339     0
#> 2003-06  0.34432844  4.02751143 4.234254     0
#> 2003-07  0.34728606  4.02561191 4.278339     0
#> 2003-08  0.33371773  4.02721175 4.248241     0
#> 2003-09  0.32765959  4.04060946 4.328283     0
#> 2003-10  0.39371467  4.04983335 4.315380     0
#> 2003-11  0.37761041  4.06245120 4.194019     0
#> 2003-12  0.33406792  4.06668605 4.323142     0
#> 2004-01  0.29283788  4.07329115 4.228605     0
#> 2004-02  0.28102974  4.07843011 4.136688     0
#> 2004-03  0.27260282  4.08797153 4.341022     0
#> 2004-04  0.35116167  4.09294223 4.312779     0
#> 2004-05  0.40253749  4.09723649 4.370950     0
#> 2004-06  0.39843714  4.09593149 4.383160     0
#> 2004-07  0.38266106  4.10067723 4.392822     0
#> 2004-08  0.40928713  4.10826073 4.328283     0
#> 2004-09  0.40610490  4.11770558 4.380730     0
#> 2004-10  0.38777489  4.13990866 4.323142     0
#> 2004-11  0.35729115  4.15294480 4.278339     0
#> 2004-12  0.29229641  4.15611159 4.363551     0
#> 2005-01  0.28720696  4.16162945 4.275640     0
#> 2005-02  0.24810930  4.16180412 4.302308     0
#> 2005-03  0.29969710  4.16442055 4.437483     0
#> 2005-04  0.32768373  4.17153740 4.402391     0
#> 2005-05  0.30619711  4.18066281 4.444356     0
#> 2005-06  0.29035350  4.18169065 4.475819     0
#> 2005-07  0.28095953  4.17593833 4.453446     0
#> 2005-08  0.30051192  4.18442639 4.451181     0
#> 2005-09  0.29672853  4.19461930 4.540081     0
#> 2005-10  0.29635684  4.21237669 4.562646     0
#> 2005-11  0.30250911  4.22630989 4.473604     0
#> 2005-12  0.29643119  4.23047674 4.576745     0
#> 2006-01  0.27997747  4.23794976 4.286392     0
#> 2006-02  0.26937807  4.24013238 4.388003     0
#> 2006-03  0.29635684  4.24287409 4.554499     0
#> 2006-04  0.27660849  4.25615322 4.499864     0
#> 2006-05  0.44708678  4.27473351 4.562646     0
#> 2006-06  0.45330244  4.27808203 4.578743     0
#> 2006-07  0.39517905  4.28652025 4.542153     0
#> 2006-08  0.37245971  4.28211649 4.514869     0
#> 2006-09  0.40593167  4.29488785 4.584713     0
#> 2006-10  0.37672264  4.30749817 4.521232     0
#> 2006-11  0.37108068  4.32032415 4.608243     0
#> 2006-12  0.34288023  4.32263181 4.600461     0
#> 2007-01  0.34847139  4.33261969 4.462456     0
#> 2007-02  0.34811844  4.33688033 4.473604     0
#> 2007-03  0.32458038  4.34600148 4.614040     0
#> 2007-04  0.31042156  4.35798719 4.558581     0
#> 2007-05  0.27748021  4.36299865 4.635014     0
#> 2007-06  0.26830810  4.36056765 4.623628     0
#> 2007-07  0.24502245  4.35323898 4.602412     0
#> 2007-08  0.25812452  4.35345529 4.592618     0
#> 2007-09  0.18871772  4.36371253 4.627438     0
#> 2007-10  0.16079995  4.38167624 4.619804     0
#> 2007-11  0.16071480  4.40097435 4.686496     0
#> 2007-12  0.15768728  4.40317200 4.584713     0
#> 2008-01  0.15768728  4.41116631 4.568713     0
#> 2008-02  0.17684827  4.42401387 4.560615     0
#> 2008-03  0.26968357  4.43357564 4.648135     0
#> 2008-04  0.25309063  4.45020640 4.631233     0
#> 2008-05  0.19028973  4.46501859 4.664756     0
#> 2008-06  0.20012054  4.46140079 4.648135     0
#> 2008-07  0.14773011  4.46714440 4.644403     0
#> 2008-08  0.16335108  4.46469611 4.556542     0
#> 2008-09  0.21070653  4.46920149 4.586695     0
#> 2008-10  0.44727861  4.49487725 4.554499     0
#> 2008-11  0.44862038  4.50316726 4.550401     0
#> 2008-12  0.42228945  4.49906201 4.392822     0
#> 2009-01  0.49375067  4.50192502 4.325716     0
#> 2009-02  0.52558615  4.49850089 4.286392     0
#> 2009-03  0.51414012  4.50947877 4.407142     0
#> 2009-04  0.45894268  4.50966380 4.423594     0
#> 2009-05  0.43061290  4.51605713 4.475819     0
#> 2009-06  0.42422144  4.51715953 4.542153     0
#> 2009-07  0.38706350  4.51966603 4.552452     0
#> 2009-08  0.40479822  4.51666973 4.491187     0
#> 2009-09  0.39241364  4.52058147 4.484629     0
#> 2009-10  0.39622253  4.54439140 4.614040     0
#> 2009-11  0.39615524  4.55699922 4.519116     0
#> 2009-12  0.39937996  4.56227907 4.608243     0
#> 2010-01  0.39655890  4.58059948 4.407142     0
#> 2010-02  0.43165254  4.59497263 4.416576     0
#> 2010-03  0.41828261  4.60078840 4.586695     0
#> 2010-04  0.39031763  4.60673846 4.556542     0
#> 2010-05  0.45126668  4.60315021 4.600461     0
#> 2010-06  0.45584132  4.59751766 4.644403     0
#> 2010-07  0.41088373  4.59270492 4.651852     0
#> 2010-08  0.42169928  4.59672657 4.614040     0
#> 2010-09  0.36942334  4.60891876 4.580737     0
#> 2010-10  0.36015389  4.62706988 4.713022     0
#> 2010-11  0.40726349  4.62734404 4.614040     0
#> 2010-12  0.43262622  4.62432414 4.809442     0
#> 2011-01  0.47187687  4.62843993 4.623628     0
#> 2011-02  0.46793900  4.63569722 4.592618     0
#> 2011-03  0.43178242  4.63987485 4.735456     0
#> 2011-04  0.41726192  4.64850024 4.697191     0
#> 2011-05  0.46354532  4.67242013 4.742259     0
#> 2011-06  0.48217921  4.65800937 4.783759     0
#> 2011-07  0.51724498  4.65390696 4.770665     0
#> 2011-08  0.55938719  4.66114142 4.733748     0
#> 2011-09  0.61826198  4.66863961 4.759066     0
#> 2011-10  0.55961579  4.70083701 4.863426     0
#> 2011-11  0.61266896  4.71795821 4.749016     0
#> 2011-12  0.63840056  4.72370033 4.888631     0
#> 2012-01  0.56999036  4.72931065 4.682905     0
#> 2012-02  0.55486165  4.73493890 4.661086     0
#> 2012-03  0.57503909  4.73906586 4.793468     0
#> 2012-04  0.56138565  4.75410228 4.750698     0
#> 2012-05  0.61239797  4.75197492 4.829835     0
#> 2012-06  0.59379635  4.74298053 4.822041     0
#> 2012-07  0.58260660  4.74068252 4.812607     0
#> 2012-08  0.59633337  4.74624731 4.702496     0
#> 2012-09  0.58165680  4.75651424 4.832936     0
#> 2012-10  0.58199213  4.77593418 4.806268     0
#> 2012-11  0.57756796  4.77970741 4.881284     0
#> 2012-12  0.57768021  4.78351336 4.884229     0
#> 2013-01  0.56477389  4.79984044 4.725165     0
#> 2013-02  0.59144663  4.80283494 4.709526     0
#> 2013-03  0.59627829  4.80943714 4.818906     0
#> 2013-04  0.58606296  4.81363250 4.832936     0
#> 2013-05  0.62476127  4.81508764 4.885699     0
#> 2013-06  0.65694989  4.82269259 4.878330     0
#> 2013-07  0.65536743  4.82579937 4.906048     0
#> 2013-08  0.72151595  4.82476485 4.716507     0
#> 2013-09  0.71098710  4.83238615 4.933020     0
#> 2013-10  0.68841098  4.85022440 4.840646     0
#> 2013-11  0.70272616  4.85031214 4.974178     0
#> 2013-12  0.75902855  4.85490764 4.979540     0
#> 2014-01  0.82487695  4.87449533 4.837569     0
#> 2014-02  0.80477800  4.87876811 4.788625     0
#> 2014-03  0.78472312  4.88996131 4.907486     0
#> 2014-04  0.75136892  4.90332137 4.897378     0
#> 2014-05  0.73883731  4.90726582 4.911787     0
#> 2014-06  0.75391297  4.91036891 4.927402     0
#> 2014-07  0.73895672  4.91490275 4.872395     0
#> 2014-08  0.77027025  4.91584815 4.858912     0
#> 2014-09  0.82459202  4.91724405 4.994138     0
#> 2014-10  0.79710171  4.93606154 4.891555     0
#> 2014-11  0.79396489  4.93787171 4.971486     0
#> 2014-12  0.84197658  4.93344104 5.052979     0
#> 2015-01  0.88366423  4.94440127 4.828281     0
#> 2015-02  0.92028275  4.95152298 4.798287     0
#> 2015-03  0.96032676  4.96330707 4.967435     0
#> 2015-04  0.98087925  4.97947469 4.963367     0
#> 2015-05  0.97873957  4.98508750 4.959282     0
#> 2015-06  0.98905781  4.97993720 5.015000     0
#> 2015-07  1.02137121  4.98082310 4.920334     0
#> 2015-08  1.07064139  4.98481911 4.986202     0
#> 2015-09  1.11384567  4.99371416 4.941388     0
#> 2015-10  1.07074422  5.00913941 5.054220     0
#> 2015-11  1.07231965  5.01577899 5.042991     0
#> 2015-12  1.06822181  5.01785877 5.127176     0
#> 2016-01  1.08638787  5.03587462 4.879808     0
#> 2016-02  1.07565068  5.03565597 4.927402     0
#> 2016-03  1.04237698  5.03525499 5.032903     0
#> 2016-04  1.03101131  5.04306341 4.998083     0
#> 2016-05  1.08473308  5.04883502 5.039220     0
#> 2016-06  1.06339953  5.05353584 5.041736     0
#> 2016-07  1.10368275  5.06510366 4.829835     0
#> 2016-08  1.08419215  5.06219761 5.023991     0
#> 2016-09  1.09814551  5.06400603 4.903167     0
#> 2016-10  1.13224047  5.07828696 5.093155     0
#> 2016-11  1.23051736  5.08344251 5.099104     0
#> 2016-12  1.25912838  5.09974338 5.132926     0
#> 2017-01  1.34439105  5.12405740 4.941388     0
#> 2017-02  1.27854210  5.13213174 4.933020     0
#> 2017-03  1.29250552  5.14224032 5.095539     0
#> 2017-04  1.26796117  5.15522730 5.067776     0
#> 2017-05  1.27185105  5.15968007 5.099104     0
#> 2017-06  1.25570150  5.15700433 5.020148     0
#> 2017-07  1.26197753  5.15848816 5.088370     0
#> 2017-08  1.23666262  5.16366427 5.099104     0
#> 2017-09  1.26843945  5.17015271 5.071441     0
#> 2017-10  1.32797645  5.19071218 5.200542     0
#> 2017-11  1.37475302  5.20552084 5.194090     0
#> 2017-12  1.32847985  5.21235517 5.234265     0
#> 2018-01  1.33049091  5.22250477 5.059171     0
#> 2018-02  1.33149493  5.22979484 5.022712     0
#> 2018-03  1.37433564  5.23968068 5.158960     0
#> 2018-04  1.40048076  5.25823081 5.109723     0
#> 2018-05  1.50127347  5.27432098 5.162307     0
#> 2018-06  1.52877001  5.30010948 5.040479     0
#> 2018-07  1.58706960  5.30560576 5.160077     0
#> 2018-08  1.85817903  5.32830056 4.979540     0
#> 2018-09  1.79102587  5.38944023 5.108549     0
#> 2018-10  1.70933754  5.41577648 5.152234     0
#> 2018-11  1.64278564  5.40124218 5.121392     0
#> 2018-12  1.66120459  5.39718826 5.129480     0
#> 2019-01  1.66446572  5.40776984 4.980876     0
#> 2019-02  1.66681019  5.40937630 4.971486     0
#> 2019-03  1.72873092  5.41960695 5.112068     0
#> 2019-04  1.78062268  5.43639628 5.097917     0
#> 2019-05  1.77983027  5.44584785 5.161192     0
#> 2019-06  1.75098088  5.44611382 4.939998     0
#> 2019-07  1.71846271  5.45958551 5.156723     0
#> 2019-08  1.76117139  5.46813582 4.962007     0
#> 2019-09  1.73416567  5.47801965 5.149982     0
#> 2019-10  1.74771179  5.49784340 5.178872     0
#> 2019-11  1.74841694  5.50164708 5.161192     0
#> 2019-12  1.78264304  5.50905241 5.224895     1
#> 2020-01  1.78791877  5.52246938 5.054220     1
#> 2020-02  1.82961707  5.52597984 5.078731     1
#> 2020-03  1.87515808  5.53167761 5.102656     1
#> 2020-04  1.94376499  5.54018585 4.725165     1
#> 2020-05  1.91984481  5.55371537 4.798287     1
#> 2020-06  1.92400775  5.56498419 5.102656     1
#> 2020-07  1.93779877  5.57080613 5.149982     1
#> 2020-08  1.98946213  5.57941249 5.112068     1
#> 2020-09  2.05605136  5.58909861 5.256807     1
#> 2020-10  2.11814529  5.61018607 5.270891     1
#> 2020-11  2.05470690  5.63290602 5.244574     1
#> 2020-12  1.99430568  5.64532409 5.340350     1
#> 2021-01  1.99172992  5.66200244 5.134072     1
#> 2021-02  1.99842805  5.67103999 5.136360     1
#> 2021-03  2.12028354  5.68173630 5.288714     1
#> 2021-04  2.10201606  5.69838678 5.231151     1
#> 2021-05  2.14679056  5.70723317 5.144329     1
#> 2021-06  2.16481616  5.72646790 5.315821     1
#> 2021-07  2.13078202  5.74435332 5.134072     1
#> 2021-08  2.12618199  5.75548675 5.286749     1
#> 2021-09  2.18055869  5.76793555 5.339418     1
#> 2021-10  2.25022281  5.79159074 5.323433     1
#> 2021-11  2.53930803  5.82610593 5.377849     1
#> 2021-12  2.56411824  5.95340348 5.462164     1
#> 2022-01  2.61145486  6.05870139 5.197321     1
#> 2022-02  2.64277541  6.10566619 5.251728     1
#> 2022-03  2.68505448  6.15886783 5.365202     1
#> 2022-04  2.69452582  6.22884484 5.319634     1
#> 2022-05  2.79550842  6.25821724 5.249689     1
#> 2022-06  2.81445317  6.30654938 5.392994     1
#> 2022-07  2.88535922  6.32992671 5.158960     1
#> 2022-08  2.90038534  6.34442605 5.325327     1
#> 2022-09  2.91887553  6.37472111 5.340350     1
#> 2022-10  2.92341692  6.40955515 5.351472     1
#> 2022-11  2.92481344  6.43798481 5.360646     1
#> 2022-12  2.92933336  6.44974225 5.470381     1
#> 2023-01  2.93409886  6.51411460 5.262867     1
#> 2023-02  2.93811375  6.54508063 5.160077     1
#> 2023-03  2.95299177  6.56771728 5.368832     1
#> 2023-04  2.96755397  6.59129228 5.215436     1
#> 2023-05  3.01203368  6.59172294 5.358818     1
#> 2023-06  3.25216939  6.63017892 5.283795     1
#> 2023-07  3.29448966  6.72083122 5.337551     1
#> 2023-08  3.28092061  6.80780487 5.358818     1
#> 2023-09  3.31059228  6.85424097 5.381433     1
#> 2023-10  3.34037639  6.88800433 5.406170     1
#> 2023-11  3.36334320  6.92028852 5.378746     1
#> 2023-12  3.38319463  6.94914034 5.464636     1
#> 2024-01  3.41222304  7.01402233 5.277860     1
#> 2024-02  3.43875325  7.05831849 5.300421     1
#> 
#> $LongCoef
#> [1] "L{lag}.{varName}"
#> 
#> $mode
#> [1] "quick"
#> 
#> $AsymPrefix
#> [1] "asyP_" "asyN_"
#> 
# Get specific options
kardl_get("maxlag", "mode")
#> $maxlag
#> [1] 2
#> 
#> $mode
#> [1] "quick"
#> 

# Note: In interactive use, avoid calling kardl_get() directly to prevent cluttering the console.

a<-kardl_get()
a$AsymSuffix
#> [1] "_PP" "_NN"

# To reset options and then get a default option:
mydefaults <- kardl_reset()
mydefaults$maxlag
#> [1] 4